+++ deployments/chart/templates/_helpers.tpl
@@ -24,10 +24,10 @@ Create chart name and version as used by the chart label.
 {{- end }}
 
 {{/*
-If replicaCount is greater than 1 leader election is enabled by default.
+If replicaCount is greater than 1, leader election is enabled by default.
 */}}
 {{- define "go-kube-downscaler.leaderElection" -}}
-{{- if (.Values.replicaCount | int | gt 1) }}
+{{- if (.Values.replicaCount | int | gt 1) }}true{{ else }}false{{- end }}
 {{- end }}
 
 {{/*
@@ -50,6 +50,14 @@ application: {{ include "go-kube-downscaler.fullname" . }}
 {{- end }}
 
 {{/*
+Selector labels for admission controller
+*/}}
+{{- define "go-kube-downscaler.admissionController.selectorLabels" -}}
+application: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+{{- end }}
+
+
+{{/*
 Create the name of the service account to use
 */}}
 {{- define "go-kube-downscaler.serviceAccountName" -}}
@@ -58,6 +66,45 @@ Create the name of the service account to use
 {{- else }}
 {{- default "default" .Values.serviceAccount.name }}
 {{- end }}
+{{- end }}
+
+{{/*
+Create the name of the service account to use for admission controller
+*/}}
+{{- define "go-kube-downscaler.admissionController.serviceAccountName" -}}
+{{- if .Values.serviceAccount.create }}
+{{- default (include "go-kube-downscaler.admissionController.fullName" .) .Values.serviceAccount.name }}
+{{- else }}
+{{- default "default" .Values.serviceAccount.name }}
+{{- end }}
+{{- end }}
+
+{{/*
+Generate the name for the admission webhook full name
+*/}}
+{{- define "go-kube-downscaler.admissionController.fullName" -}}
+{{- printf "%s-admission-webhook" (include "go-kube-downscaler.fullname" .) }}
+{{- end }}
+
+{{/*
+Generate the name for the admission webhook secret
+*/}}
+{{- define "go-kube-downscaler.admissionController.secretName" -}}
+{{- printf "%s-secret" (include "go-kube-downscaler.admissionController.fullName" .) }}
+{{- end }}
+
+{{/*
+Create defined permissions for admission webhook deployment
+*/}}
+{{- define "go-kube-downscaler.admissionController.permissions" -}}
+- apiGroups:
+    - ""
+  resources:
+    - secrets
+  verbs:
+    - get
+    - watch
+    - list
 {{- end }}
 
 {{/*
@@ -271,6 +318,135 @@ Create defined permissions for roles
     - list
     - update
     - patch
+{{- end }}
+{{- end }}
+{{- end }}
+
+{{/*
+Create webhook resources
+*/}}
+{{- define "go-kube-downscaler.webhookresources" -}}
+{{- range $resource := .Values.includedResources }}
+{{- if eq $resource "deployments" }}
+- apiGroups:
+    - apps
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - deployments
+{{- end }}
+{{- if eq $resource "statefulsets" }}
+- apiGroups:
+    - apps
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - statefulsets
+{{- end }}
+{{- if eq $resource "daemonsets" }}
+- apiGroups:
+    - apps
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - daemonsets
+{{- end }}
+{{- if eq $resource "rollouts" }}
+- apiGroups:
+    - argoproj.io
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - rollouts
+{{- end }}
+{{- if eq $resource "horizontalpodautoscalers" }}
+- apiGroups:
+    - autoscaling
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - horizontalpodautoscalers
+{{- end }}
+{{- if eq $resource "jobs" }}
+- apiGroups:
+    - batch
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - jobs
+{{- end }}
+{{- if eq $resource "cronjobs" }}
+- apiGroups:
+    - batch
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - cronjobs
+{{- end }}
+{{- if eq $resource "scaledobjects" }}
+- apiGroups:
+    - keda.sh
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - scaledobjects
+{{- end }}
+{{- if eq $resource "stacks" }}
+- apiGroups:
+    - zalando.org
+  resources:
+    - stacks
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+{{- end }}
+{{- if eq $resource "prometheuses" }}
+- apiGroups:
+    - monitoring.coreos.com
+  resources:
+    - prometheuses
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+{{- end }}
+{{- if eq $resource "poddisruptionbudgets" }}
+- apiGroups:
+    - policy
+  apiVersions:
+    - "*"
+  operations:
+    - "CREATE"
+    - "UPDATE"
+  resources:
+    - poddisruptionbudgets
 {{- end }}
 {{- end }}
 {{- end }}
