+++ deployments/chart/templates/admissionrole.yaml
@@ -1,0 +1,23 @@
+{{- if .Values.admissionController.enabled }}
+apiVersion: rbac.authorization.k8s.io/v1
+kind: Role
+metadata:
+  name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+  namespace: {{ .Release.Namespace }}
+rules:
+{{ include "go-kube-downscaler.admissionController.permissions" . }}
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+  namespace: {{ .Release.Namespace }}
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: Role
+  name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+subjects:
+  - kind: ServiceAccount
+    name: {{ include "go-kube-downscaler.admissionController.serviceAccountName" . }}
+    namespace: {{ .Release.Namespace }}
+{{- end }}
