+++ deployments/chart/templates/admissionsecret.yaml
@@ -1,0 +1,10 @@
+{{- if .Values.admissionController.enabled }}
+apiVersion: v1
+data:
+  tls.crt: {{ .Values.admissionController.tlsCrt }}
+  tls.key: {{ .Values.admissionController.tlsKey }}
+kind: Secret
+metadata:
+  name: {{ include "go-kube-downscaler.admissionController.secretName" . }}
+type: kubernetes.io/tls
+{{- end }}
