+++ deployments/chart/templates/admissionserviceaccount.yaml
@@ -1,0 +1,8 @@
+{{- if .Values.admissionController.enabled }}
+{{- if  .Values.admissionController.serviceAccount.create}}
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ include "go-kube-downscaler.admissionController.serviceAccountName" . }}
+{{- end }}
+{{- end }}
