+++ deployments/chart/templates/validatingwebhookconfiguration.yaml
@@ -1,0 +1,20 @@
+{{- if .Values.admissionController.enabled }}
+apiVersion: admissionregistration.k8s.io/v1
+kind: ValidatingWebhookConfiguration
+metadata:
+  name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+webhooks:
+  - name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+    rules:
+{{ include "go-kube-downscaler.webhookresources" . | trim | nindent 6 }}
+    clientConfig:
+      service:
+        namespace: {{ .Release.Namespace }}
+        name: {{ include "go-kube-downscaler.admissionController.fullName" . }}
+        path: /validate-workloads
+        port: 443
+      caBundle: {{ .Values.admissionController.caBundle }}
+    admissionReviewVersions: ["v1"]
+    sideEffects: None
+    timeoutSeconds: 2
+{{- end }}
