+++ internal/api/kubernetes/admission/admissionHandler.go
@@ -1,0 +1,81 @@
+package admission
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	admissionv1 "k8s.io/api/admission/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"log/slog"
+	"net/http"
+)
+
+type admissionHandler interface{
+	// HandleValidation validates the admission request and returns an AdmissionResponse
+	HandleValidation() *admissionv1.AdmissionResponse
+}
+
+// parseAdmissionReviewFromRequest extracts an AdmissionReview from a http.Request if possible
+func parseAdmissionReviewFromRequest(r http.Request) (*admissionv1.AdmissionReview, error) {
+	if r.Header.Get("Content-Type") != "application/json" {
+		return nil, fmt.Errorf("Content-Type: %q should be %q",
+			r.Header.Get("Content-Type"), "application/json")
+	}
+
+	bodybuf := new(bytes.Buffer)
+	_, err := bodybuf.ReadFrom(r.Body)
+	if err != nil {
+		return nil, err
+	}
+	body := bodybuf.Bytes()
+
+	if len(body) == 0 {
+		return nil, fmt.Errorf("admission request body is empty")
+	}
+
+	var a admissionv1.AdmissionReview
+
+	if err := json.Unmarshal(body, &a); err != nil {
+		return nil, fmt.Errorf("could not parse admission review request: %v", err)
+	}
+
+	if a.Request == nil {
+		return nil, fmt.Errorf("admission review can't be used: Request field is nil")
+	}
+
+	return &a, nil
+}
+
+// reviewResponse returns an AdmissionReview with the specified UID, allowed, httpCode, and reason
+func reviewResponse(uid types.UID, allowed bool, httpCode int32, reason string) *admissionv1.AdmissionReview {
+	return &admissionv1.AdmissionReview{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "AdmissionReview",
+			APIVersion: "admission.k8s.io/v1",
+		},
+		Response: &admissionv1.AdmissionResponse{
+			UID:     uid,
+			Allowed: allowed,
+			Result: &metav1.Status{
+				Code:    httpCode,
+				Message: reason,
+			},
+		},
+	}
+}
+
+// sendAdmissionResponse sends the admission response to the client
+func sendAdmissionReviewResponse(w http.ResponseWriter, err error, out *admissionv1.AdmissionReview) {
+	w.Header().Set("Content-Type", "application/json")
+	jout, err := json.Marshal(out)
+	if err != nil {
+		e := fmt.Sprintf("could not parse admission response: %s", err)
+		slog.Error(e)
+		http.Error(w, e, http.StatusInternalServerError)
+		return
+	}
+
+	slog.Debug("sending response")
+	slog.Debug("%s", jout)
+}
